add spring security and change user entity to userInfo, and add the id-pw info in new User entity

+++ src/main/java/com/instapic/coreservice/dto/response/user/UserDetailResponseDto.java
@@ -15,8 +15,8 @@ public class UserDetailResponseDto {
     private String profilePictureUrl;
     private String bio;
     private String url;
-    private Long followingCount;
-    private Long followedByCount;
+    private int followingCount;
+    private int followedByCount;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
